UVEditor: Clean up warnings related to type truncation and loss of precision issues.
#rb michael.balzer
#preflight 6331f24ee11ae0b19cd5480d
#jira UE-165106

[CL 22230411 by nathan mitchell in ue5-release-engine-staging branch]

+++ Engine/Plugins/Experimental/UVEditor/Source/UVEditor/Private/UVEditor2DViewportBehaviorTargets.cpp
@@ -16,7 +16,7 @@ FInputRayHit FUVEditor2DScrollBehaviorTarget::CanBeginClickDragSequence(const FI
 	// Verify that ray is facing the proper direction
 	if (PressPos.WorldRay.Direction.Z * PressPos.WorldRay.Origin.Z < 0)
 	{
-		return FInputRayHit(-PressPos.WorldRay.Origin.Z / PressPos.WorldRay.Direction.Z);
+		return FInputRayHit(static_cast<float>(-PressPos.WorldRay.Origin.Z / PressPos.WorldRay.Direction.Z));
 	}
 	return FInputRayHit();
 }
@@ -90,8 +90,8 @@ void FUVEditor2DMouseWheelZoomBehaviorTarget::OnMouseWheelScrollUp(const FInputD
 
 	FVector NewLocation = OriginalLocation + (ZoomInProportion * TToPlane * CurrentPos.WorldRay.Direction);
 
-	ViewportClient->OverrideFarClipPlane(NewLocation.Z - CameraFarPlaneWorldZ);
-	ViewportClient->OverrideNearClipPlane(NewLocation.Z * (1.0-CameraNearPlaneProportionZ));
+	ViewportClient->OverrideFarClipPlane(static_cast<float>(NewLocation.Z - CameraFarPlaneWorldZ));
+	ViewportClient->OverrideNearClipPlane(static_cast<float>(NewLocation.Z * (1.0-CameraNearPlaneProportionZ)));
 
 	// Don't zoom in so far that the XY plane lies in front of our near clipping plane, or else everything
 	// will suddenly disappear.
@@ -107,8 +107,8 @@ void FUVEditor2DMouseWheelZoomBehaviorTarget::OnMouseWheelScrollDown(const FInpu
 	double TToPlane = -OriginalLocation.Z / CurrentPos.WorldRay.Direction.Z;
 	FVector NewLocation = OriginalLocation - (ZoomOutProportion * TToPlane * CurrentPos.WorldRay.Direction);
 
-	ViewportClient->OverrideFarClipPlane(NewLocation.Z - CameraFarPlaneWorldZ);
-	ViewportClient->OverrideNearClipPlane(NewLocation.Z * (1.0 - CameraNearPlaneProportionZ));
+	ViewportClient->OverrideFarClipPlane(static_cast<float>(NewLocation.Z - CameraFarPlaneWorldZ));
+	ViewportClient->OverrideNearClipPlane(static_cast<float>(NewLocation.Z * (1.0 - CameraNearPlaneProportionZ)));
 
 	if (NewLocation.Z < ZoomOutLimit)
 	{
